Removed support for executing benchmarks until predefined confidence level is reached
This does not seem to be sound methodology, because it disregards the underlying reason for the differences in the measurements. When executing until a certain confidence level is reached, the experiment is just run as many times as necessary to reach a desired result.
Better methodology is to define upfront how many samples/measurements are to be taken and then to assess the statistic properties of this set of samples.
This also makes multiple different experiments more comparable, because the number of measurements is a constant.

This closes #14.

The ReBench configuration files changed.
It has now a section `runs:` which defines the number of data points to be taken, and optionally a minimum value for the measure runtime which will cause a warning if violated. (The warning is currently disabled).

For reporting purposes, a confidence level can be set in the `reporting:` section, which is used to calculate the confidence interval.

The configuration for quick runs was changed accordingly.

TODO:
 - [ ] report on change
 - [ ] reenable support for min_runtime warning

Signed-off-by: Stefan Marr <git@stefan-marr.de>