Added support to define a command in the benchmark definition.
NOTE: untested...

It does not supersede codespeed_name, because there might still be reason to keep them separate. For instance, to use identical names in the ReBench config and data files for different suites, but provide codespeed with unique names. In the data processing that might not be necessary, because the data contains the suite name.

This closes issue #28.

Signed-off-by: Stefan Marr <git@stefan-marr.de>