Add queue-based pulling parallel scheduler
The existing parallel scheduler behaves as is
(With common behavior lifted to a new base clase).

A new parallel scheulder maintains a work queue and the workers retrieve
directly from there.

a distinction is made for py2 vs py3, to make use of the built-in thread
pool executor when possible.

No new magic so far