A build command is now built by combining all commands in the list to a single script
The main reason for this change is to make the deduplication more sensible.

Previously, something like a `make clean` could have been deduplicated even if the commands are somewhat unrelated but accidentally split into separate lines. The semantics around this were not particularly intuitive.
By combining everything into a single command it becomes a little more predictable. And overall simpler.

- make meaning of deduplicated_build_commands field explicit by naming it so
- the first PR that adds type annotations, I believe

Signed-off-by: Stefan Marr <git@stefan-marr.de>