Base RunId comparison and hash based on all details
This is in preparation for the support of #257, which will add machine and denoise configuration support.
Though, it’s also useful without, because it allows us to distinguish RunIds that differ in their environment variables, for instance. Or more generally, RunIds that differ based on any property that is not strictly part of the command line, which is was previously used to establish equality.

This as a consequence means, we have a much weaker ability to determine equality of RunIds than before, but I think that’s fine and less surprising/buggy.

The main changes are a proper implementation of the __eq__, __lt__, and __hash__ methods, as well as the serialization of RunIds into the data file, and deserialization when loading a data file.

This change includes some minor refactorings, because it is split from a too huge and unmanagable patch.

Minor refactorings:
 - configurator: extract config validation and assembling of run details
 - executor: use exe name and suite name directly in indicate_build
 - executor/BuildCommand: location is in BuildCommand only for equality, but semantically, we should use the location/path of the suite/executor when processing the build command. That’s now made explicit, and asserted for correctness.
 - BuildCommand is now storing only the command, since the location is used from suite/exe

Signed-off-by: Stefan Marr <git@stefan-marr.de>